service_analyze/requirements.txt
python-pptx
python-docx
python-dateutil
pdf2docx
groq
chardet
beautifulsoup4
lxml
html5lib
PyPDF2
html2text
openpyxl
pandas
Flask
requests
# chroma
# sentence_transformers